+++ spinetoolbox/widgets/data_store_subwindow_widget.py
@@ -24,12 +24,8 @@ QWidget that is used to display information contained in a Data Store.
 :date:   19.4.2018
 """
 
-import logging
-from PySide2.QtGui import QStandardItemModel, QStandardItem, QIcon, QPixmap
 from PySide2.QtWidgets import QWidget
-from PySide2.QtCore import Qt
 from ui.subwindow_data_store import Ui_Form
-from config import DC_TREEVIEW_HEADER_SS, HEADER_POINTSIZE
 
 
 class DataStoreWidget(QWidget):
@@ -46,14 +42,6 @@ class DataStoreWidget(QWidget):
         self.ui = Ui_Form()
         self.ui.setupUi(self)
         self.setObjectName(item_type)  # TODO: Remove. item_type is an instance variable of DataStore objects
-        self.reference_model = QStandardItemModel()  # References to databases
-        self.data_model = QStandardItemModel()  # Paths of project internal Spine objects. Found in DS data directory.
-        self.spine_icon = QIcon(QPixmap(":/icons/Spine_db_icon.png"))
-        self.spine_ref_icon = QIcon(QPixmap(":/icons/Spine_db_ref_icon.png"))
-        self.ui.treeView_references.setModel(self.reference_model)
-        self.ui.treeView_data.setModel(self.data_model)
-        self.ui.treeView_references.setStyleSheet(DC_TREEVIEW_HEADER_SS)
-        self.ui.treeView_data.setStyleSheet(DC_TREEVIEW_HEADER_SS)
         self.ui.label_name.setFocus()
 
     def owner(self):
@@ -71,53 +59,6 @@ class DataStoreWidget(QWidget):
     def name_label(self):
         """Return name label text."""
         return self.ui.label_name.text()
-
-    def make_header_for_references(self):
-        """Add header to files model. I.e. External Data Connection files."""
-        h = QStandardItem("References")
-        # Decrease font size
-        font = h.font()
-        font.setPointSize(HEADER_POINTSIZE)
-        h.setFont(font)
-        self.reference_model.setHorizontalHeaderItem(0, h)
-
-    def make_header_for_data(self):
-        """Add header to data model. I.e. Internal Data Connection files."""
-        h = QStandardItem("Data")
-        # Decrease font size
-        font = h.font()
-        font.setPointSize(HEADER_POINTSIZE)
-        h.setFont(font)
-        self.data_model.setHorizontalHeaderItem(0, h)
-
-    def populate_reference_list(self, items):
-        """List file references in QTreeView.
-        If items is None or empty list, model is cleared.
-        """
-        self.reference_model.clear()
-        self.make_header_for_references()
-        if items is not None:
-            for item in items:
-                qitem = QStandardItem(item['database'])
-                qitem.setFlags(~Qt.ItemIsEditable)
-                qitem.setData(item['url'], Qt.ToolTipRole)
-                qitem.setData(self.spine_ref_icon, Qt.DecorationRole)
-                self.reference_model.appendRow(qitem)
-
-    def populate_data_list(self, items):
-        """List project internal data (files) in QTreeView.
-        If items is None or empty list, model is cleared.
-        """
-        self.data_model.clear()
-        self.make_header_for_data()
-        if items is not None:
-            for item in items:
-                qitem = QStandardItem(item)
-                qitem.setFlags(~Qt.ItemIsEditable)
-                qitem.setData(item, Qt.ToolTipRole)
-                if item.endswith('sqlite'):
-                    qitem.setData(self.spine_icon, Qt.DecorationRole)
-                self.data_model.appendRow(qitem)
 
     def closeEvent(self, event):
         """Hide widget and is proxy instead of closing them.